+++ src/DB/achieveTable.ts
@@ -1,5 +1,6 @@
 import Dexie from "dexie";
-import { Achieve } from "../components/Achievement";
+import { AchievePage } from "../components/Achievement";
+import { Page } from "./db";
 
 export type Achievement = {
   id?: number;
@@ -10,25 +11,32 @@ export type Achievement = {
   description: string;
 };
 
-// レコード追加
-async function create(table: Dexie.Table<any, number>, object: object) {
-  await table.put(object);
-  table.toCollection().toArray();
+// レコード追加、IDを返す
+async function create(
+  table: Dexie.Table<Achievement, number>,
+  object: Achievement
+): Promise<number> {
+  const newRecordID = await table.put(object);
+  return newRecordID;
 }
 
 // idからレコード検索
-function find(table: Dexie.Table<any, number>, id: number) {
-  return table.get(id, res => res);
+function find(table: Dexie.Table<Achievement, number>, id: number) {
+  const res = table.get(id);
+  if (typeof res === "undefined") {
+    throw new Error("undifined");
+  }
+  return res;
 }
 
-// AchievePage
+// Achieve一覧ページ
 async function createPage(
-  pageTable: Dexie.Table<any, number>,
-  achieveTable: Dexie.Table<any, number>
-): Dexie.Promise<Achieve> {
+  pageTable: Dexie.Table<Page, number>,
+  achieveTable: Dexie.Table<Achievement, number>
+): Dexie.Promise<AchievePage> {
   const page = await pageTable.get(2, res => res);
   const achievements = await achieveTable.toCollection().toArray();
-  return { ...page, achievements };
+  return { ...page, achievements } as AchievePage;
 }
 
 export default { create, find, createPage };
