+++ src/DB/reviewTable.ts
@@ -1,5 +1,6 @@
 import Dexie from "dexie";
 import { ReviewPage } from "../components/Review";
+import { Page } from "./db";
 
 export interface Review {
   id?: number;
@@ -11,23 +12,31 @@ export interface Review {
 }
 
 // レコード追加
-async function create(table: Dexie.Table<any, number>, object: object) {
-  await table.put(object);
+async function create(table: Dexie.Table<Review, number>, object: Review) {
+  try {
+    return await table.add(object);
+  } catch (err) {
+    console.error(err);
+  }
 }
 
 // idからレコード検索
-function find(table: Dexie.Table<any, number>, id: number) {
-  return table.get(id, res => res);
+async function find(table: Dexie.Table<Review, number>, id: number) {
+  const res = await table.get(id);
+  if (!res) {
+    throw new Error("Not find records");
+  }
+  return res;
 }
 
 // ReviewPageを生成する
 async function createPage(
-  pageTable: Dexie.Table<any, number>,
-  reviewTable: Dexie.Table<any, number>
+  pageTable: Dexie.Table<Page, number>,
+  reviewTable: Dexie.Table<Review, number>
 ): Dexie.Promise<ReviewPage> {
   const page = await pageTable.get(1, res => res);
   const reviews = await reviewTable.toCollection().toArray();
-  return { ...page, reviews };
+  return { ...page, reviews } as ReviewPage;
 }
 
 export default { create, find, createPage };
